switch to US layout with a few changes
I now remap my keys to the NEO layout in the operating system, not in
the keyboard. This unifies my setup across computers and makes sure I
use the same variant of my layout everywhere (as opposed to an Xmodmap
file which might go out of sync with the XKB version).

The only remaining two hardware-level modifications are:

• Delete and Escape are swapped
• “Home” is replaced by the left windows key (Super_L, modifier)